requirements.txt
streamlit
pandas
google-api-python-client
google-auth
google-auth-oauthlib
spacy==3.7.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz